pixie-backend/pyproject.toml
[project]
name = "pixie-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.124.4",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.8",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.7",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.27.0",
]
